+++ hikka/inline/bot_pm.py
@@ -43,24 +43,3 @@ class BotPM(InlineUnit):
         return self.fsm.get(str(user), False)
 
     gs = get_fsm_state
-
-    async def _bot_message_answer(  # skipcq: PYL-E0213
-        mod,
-        text: str = None,
-        message: AiogramMessage = None,
-        parse_mode: Optional[str] = "HTML",
-        disable_web_page_preview: Optional[bool] = True,
-        **kwargs,
-    ) -> bool:
-        try:
-            await mod.bot.send_message(
-                message.chat.id,
-                text,
-                parse_mode=parse_mode,
-                disable_web_page_preview=disable_web_page_preview,
-                **kwargs,
-            )
-        except Exception:
-            return False
-
-        return True